+++ app/controllers/api/availabilities_controller.rb
@@ -147,7 +147,7 @@ class API::AvailabilitiesController < API::ApiController
 
   def availability_params
     params.require(:availability).permit(:start_at, :end_at, :available_type, :machine_ids, :training_ids, :nb_total_places,
-                                         :is_recurrent, :period, :nb_periods, :end_date,
+                                         :is_recurrent, :period, :nb_periods, :end_date, :slot_duration,
                                          machine_ids: [], training_ids: [], space_ids: [], tag_ids: [], plan_ids: [],
                                          machines_attributes: %i[id _destroy], plans_attributes: %i[id _destroy])
   end