+++ app/controllers/api/slots_controller.rb
@@ -1,7 +1,8 @@
 # frozen_string_literal: true
 
 # API Controller for resources of type Slot
-# Slots are used to cut Availabilities into reservable slots of ApplicationHelper::SLOT_DURATION minutes
+# Slots are used to cut Availabilities into reservable slots. The duration of these slots is configured per
+# availability by Availability.slot_duration, or otherwise globally by ApplicationHelper::SLOT_DURATION minutes
 class API::SlotsController < API::ApiController
   before_action :authenticate_user!
   before_action :set_slot, only: %i[update cancel]