+++ db/migrate/20200415141809_add_slot_duration_to_availability.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal:true
+
+# From this migration any availability can override the default SLOT_DURATION value for its own slots
+class AddSlotDurationToAvailability < ActiveRecord::Migration[5.2]
+  def change
+    add_column :availabilities, :slot_duration, :integer
+  end
+end